refactor: structure code consistently
