--- libs/react/containers/src/blog/blog-content/index.ts
@@ -1,1 +1,0 @@
-export { DocContent } from "./doc-content";
