+++ libs/react/containers/src/blog/blog-post/index.ts
@@ -1,0 +1,1 @@
+export { BlogPost } from "./blog-post";
