+++ libs/react/containers/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./annexes/annex-search";
+export * from "./blog/blog-post";
 export * from "./docs/doc-content";
 export * from "./docs/doc-search";
 export * from "./flashcards/flashcard-collection-list";
@@ -22,4 +23,3 @@ export * from "./tests/test-review";
 export * from "./tests/test-search";
 export * from "./tests/test-study";
 export * from "./user/user-settings";
-export * from "./user/hooks/use-user-voyage";
